# services/@tables/service/spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tables Service API",
    "description": "Dynamic data tables API with Yjs CRDT real-time collaboration",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/tables": {
      "post": {
        "operationId": "createTable",
        "tags": [
          "tables"
        ],
        "description": "Create a new table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Table name",
                    "minLength": 1
                  },
                  "columns": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "name",
                        "dataType"
                      ],
                      "properties": {
                        "name": {
                          "type": "string",
                          "description": "Column name",
                          "minLength": 1
                        },
                        "dataType": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "text"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "number"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "boolean"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "date"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "select"
                              ]
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "userId",
                    "name",
                    "columns",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Table ID (UUID)"
                    },
                    "userId": {
                      "type": "string",
                      "description": "Owner user ID"
                    },
                    "name": {
                      "type": "string",
                      "description": "Table name"
                    },
                    "columns": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "dataType",
                          "position"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Column ID (UUID)"
                          },
                          "name": {
                            "type": "string",
                            "description": "Column display name"
                          },
                          "dataType": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "text"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "number"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "boolean"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "date"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "select"
                                ]
                              }
                            ]
                          },
                          "position": {
                            "type": "number",
                            "description": "Column position (0-indexed)"
                          }
                        }
                      },
                      "description": "Table columns"
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Creation timestamp (ISO 8601)"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "Last update timestamp (ISO 8601)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listTables",
        "tags": [
          "tables"
        ],
        "description": "List all tables for the current user",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "createdAt",
                      "updatedAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Table ID (UUID)"
                      },
                      "name": {
                        "type": "string",
                        "description": "Table name"
                      },
                      "createdAt": {
                        "type": "string",
                        "description": "Creation timestamp (ISO 8601)"
                      },
                      "updatedAt": {
                        "type": "string",
                        "description": "Last update timestamp (ISO 8601)"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tables/{tableId}": {
      "get": {
        "operationId": "getTable",
        "tags": [
          "tables"
        ],
        "description": "Get a table by ID with its schema",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "userId",
                    "name",
                    "columns",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Table ID (UUID)"
                    },
                    "userId": {
                      "type": "string",
                      "description": "Owner user ID"
                    },
                    "name": {
                      "type": "string",
                      "description": "Table name"
                    },
                    "columns": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "dataType",
                          "position"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Column ID (UUID)"
                          },
                          "name": {
                            "type": "string",
                            "description": "Column display name"
                          },
                          "dataType": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "text"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "number"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "boolean"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "date"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "select"
                                ]
                              }
                            ]
                          },
                          "position": {
                            "type": "number",
                            "description": "Column position (0-indexed)"
                          }
                        }
                      },
                      "description": "Table columns"
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Creation timestamp (ISO 8601)"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "Last update timestamp (ISO 8601)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateTable",
        "tags": [
          "tables"
        ],
        "description": "Update a table (rename)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "New table name",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "userId",
                    "name",
                    "columns",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Table ID (UUID)"
                    },
                    "userId": {
                      "type": "string",
                      "description": "Owner user ID"
                    },
                    "name": {
                      "type": "string",
                      "description": "Table name"
                    },
                    "columns": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "dataType",
                          "position"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Column ID (UUID)"
                          },
                          "name": {
                            "type": "string",
                            "description": "Column display name"
                          },
                          "dataType": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "text"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "number"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "boolean"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "date"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "select"
                                ]
                              }
                            ]
                          },
                          "position": {
                            "type": "number",
                            "description": "Column position (0-indexed)"
                          }
                        }
                      },
                      "description": "Table columns"
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Creation timestamp (ISO 8601)"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "Last update timestamp (ISO 8601)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTable",
        "tags": [
          "tables"
        ],
        "description": "Delete a table",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tables/{tableId}/columns": {
      "post": {
        "operationId": "addColumn",
        "tags": [
          "columns"
        ],
        "description": "Add a new column to a table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "dataType"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Column name",
                    "minLength": 1
                  },
                  "dataType": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "text"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "number"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "boolean"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "date"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "select"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "dataType",
                    "position"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Column ID (UUID)"
                    },
                    "name": {
                      "type": "string",
                      "description": "Column display name"
                    },
                    "dataType": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "text"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "number"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "boolean"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "date"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "select"
                          ]
                        }
                      ]
                    },
                    "position": {
                      "type": "number",
                      "description": "Column position (0-indexed)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ]
}